+++ backend/app/models/sortie_fonds.py
@@ -3,6 +3,8 @@ from __future__ import annotations
 import uuid
 from datetime import datetime, timezone
 
+from decimal import Decimal
+
 from sqlalchemy import DateTime, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
@@ -22,7 +24,7 @@ class SortieFonds(Base):
     requisition_id: Mapped[uuid.UUID | None] = mapped_column(UUID(as_uuid=True), nullable=True, index=True)
     rubrique_code: Mapped[str | None] = mapped_column(String(50), nullable=True)
 
-    montant_paye: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False, default=0)
+    montant_paye: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False, default=0)
     date_paiement: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     mode_paiement: Mapped[str] = mapped_column(String(50), nullable=False)
     reference: Mapped[str | None] = mapped_column(String(100), nullable=True)
